Added support for median calculation

+++ src/main/scala/com/clicktale/performance/RunnerConfigs.scala
@@ -2,8 +2,8 @@ package com.clicktale.performance
 
 import com.typesafe.config.ConfigFactory
 
-case class RunnerConfigs(numOfBins:Long)
+case class RunnerConfigs(numOfBins:Int)
 object RunnerConfigs {
   private val runnerConfs = ConfigFactory.load().getConfig("aerospike_performance.runner")
-  def apply() = new RunnerConfigs(numOfBins = runnerConfs.getLong("numofbins"))
+  def apply() = new RunnerConfigs(numOfBins = runnerConfs.getInt("numofbins"))
 }
